feat: 예외처리 완성

+++ src/main/java/BCSD/MusicStream/repository/LyricsRepository.java
@@ -3,6 +3,6 @@ package BCSD.MusicStream.repository;
 import BCSD.MusicStream.domain.Lyrics;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface LyricsRepository extends JpaRepository<Lyrics, Long> {
+public interface LyricsRepository extends JpaRepository<Lyrics, Integer> {
 
 }
