+++ src/main/java/BCSD/MusicStream/repository/PlaylistMusicRepository.java
@@ -5,6 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.List;
 
-public interface PlaylistMusicRepository  extends JpaRepository<PlaylistMusic, Long> {
+public interface PlaylistMusicRepository  extends JpaRepository<PlaylistMusic, Integer> {
     public List<PlaylistMusic> findAllByPlaylistId(Integer playlistId);
 }
