+++ src/main/java/BCSD/MusicStream/repository/WeatherRepository.java
@@ -6,6 +6,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.Optional;
 
-public interface WeatherRepository extends JpaRepository<Weather, Long> {
+public interface WeatherRepository extends JpaRepository<Weather, Integer> {
     Optional<Weather> findWeatherByName(String name);
 }
